+++ app/src/main/java/com/github/eylulnc/walkmunich/feature/home/data/city/repository/CityRepository.kt
@@ -1,7 +1,7 @@
-package com.github.eylulnc.walkmunich.feature.main.data.city.repository
+package com.github.eylulnc.walkmunich.feature.home.data.city.repository
 
 import com.github.eylulnc.walkmunich.core.data.model.City
-import com.github.eylulnc.walkmunich.feature.main.data.city.service.CityService
+import com.github.eylulnc.walkmunich.feature.home.data.city.service.CityService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
