+++ app/src/main/java/com/github/eylulnc/walkmunich/feature/home/data/city/service/CityService.kt
@@ -1,4 +1,4 @@
-package com.github.eylulnc.walkmunich.feature.main.data.city.service
+package com.github.eylulnc.walkmunich.feature.home.data.city.service
 
 import com.github.eylulnc.walkmunich.core.data.model.City
