+++ app/src/main/java/com/github/eylulnc/walkmunich/feature/home/ui/SearchResultsSection.kt
@@ -1,6 +1,5 @@
-package com.github.eylulnc.walkmunich.feature.main.ui
+package com.github.eylulnc.walkmunich.feature.home.ui
 
-import android.content.res.Resources
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
