+++ app/src/main/java/com/github/eylulnc/walkmunich/feature/home/viewModel/HomeScreenViewModel.kt
@@ -1,12 +1,12 @@
-package com.github.eylulnc.walkmunich.feature.main.viewModel
+package com.github.eylulnc.walkmunich.feature.home.viewModel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.eylulnc.walkmunich.core.data.model.City
 import com.github.eylulnc.walkmunich.core.data.model.Place
 import com.github.eylulnc.walkmunich.core.data.model.SearchResult
-import com.github.eylulnc.walkmunich.feature.main.data.city.repository.CityRepository
-import com.github.eylulnc.walkmunich.feature.main.data.place.repository.PlacesRepository
+import com.github.eylulnc.walkmunich.feature.home.data.city.repository.CityRepository
+import com.github.eylulnc.walkmunich.feature.home.data.place.repository.PlacesRepository
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
