+++ app/src/main/java/com/github/eylulnc/walkmunich/navigation/NavigationRoot.kt
@@ -25,7 +25,7 @@ import androidx.navigation3.runtime.rememberNavBackStack
 import androidx.navigation3.runtime.rememberSavedStateNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
 import androidx.navigation3.ui.rememberSceneSetupNavEntryDecorator
-import com.github.eylulnc.walkmunich.feature.main.ui.MainScreenUi
+import com.github.eylulnc.walkmunich.feature.home.ui.MainScreenUi
 import com.github.eylulnc.walkmunich.feature.route.ui.RouteListScreenUi
 import com.github.eylulnc.walkmunich.core.ui.theme.OrangeMain
 import com.github.eylulnc.walkmunich.core.ui.theme.Spacing
